MERLS-Backend-Python-main/requirements.txt
flask >= 2.0.0
supabase >= 1.0.0
boto3 >= 1.24.0
psycopg2-binary >= 2.9.0
pandas >= 1.5.0
python-dotenv >= 1.0.0
